Transform the output of sourmash taxonomy annotate into a taxmap metacoder object

Transform the output of sourmash taxonomy annotate into a taxmap metacoder object

Usage

from_taxonomy_annotate_to_metacoder(
  taxonomy_annotate_df = NULL,
  file = NULL,
  intersect_bp_threshold = 50000,
  tax_glom_level = NULL,
  groups = NULL,
  groups_prefix = "x",
  class_key = NULL,
  class_regex = NULL
)

Arguments

taxonomy_annotate_df

Data frame containing outputs from sourmash taxonomy annotate. If specified, file is ignored. Can contain results from one or many runs of sourmash taxonomy annotate.

file

Path to CSV file or files output by sourmash taxonomy annotate.

intersect_bp_threshold

Integer. Gather matches must have an intersect_bp greater than or equal to this value.

tax_glom_level

Character. NULL by default, meaning no agglomeration is done. Valid options are "domain", "phylum", "class", "order", "family", "genus", and "species". When a valid option is supplied, k-mer counts are agglomerated to that level before metacoder object is created.

groups

A data frame with distinct query_name values from taxonomy_annotate_df in the first column and query groups in the second column.

groups_prefix

Character. Ignored if groups is defined. Used to prefix query_name values for the metacoder function calc_n_samples().

class_key

Character. class_key to use to build metacoder object. See metacoder documentation for more information. Ignored if you used pre-built sourmash GenBank or GTDB databases when running sourmash.

class_regex

Character. class_regex to use to build metacoder object. See metacoder documentation for more information. Ignored if you used pre-built sourmash GenBank or GTDB databases when running sourmash.

Value

A metacoder taxmap object.

Examples

if (FALSE) {
from_taxonomy_annotate_to_metacoder()
}