Gut microbiome sourmash signatures data frame

Sourmash sketches from six stool microbiome shotgun metagenome samples parsed into a data frame.

Usage

gut_signatures_df

Format

## `gut_signatures_df` A data frame with 15,339 rows and 16 columns:

class

Character. Class of the sketch.

email

Character. Email recorded in the sketch.

hash_function

Character. Hash function used for the sketch.

filename

Character. Name of the file that was sketched.

name

Character. Name of the sketch.

license

Character. License for the sketch.

num

Integer. Number of hashes specified to be in the sketch.

ksize

Integer. K-mer size represented in the sketch.

seed

Integer. Seed for the hash function.

max_hash

Numeric. Maximum hash.

scaled

Numeric. Scaled value for the sketch.

mins

Numeric. MinHash, or hash, that represents each k-mer sketched from the sample.

md5sum

Character. MD5 hash for the sketch.

abundances

Integer. Abundance of hash in sample.

molecule

Character. Molecule type represented by sketch.

version

Numeric. Sketch version.

...

Source

<https://github.com/Arcadia-Science/sourmashconsumr/blob/main/data-raw/00_sourmash_commands.sh>